Cleanup FileUtil
- Moving detection of Temp-Folder into FileUtil
- Vaporizing BasicFileUtil and safe-mode detection
- Vaporizing whitespace at EOL
