+++ wcfsetup/install/files/lib/system/WCF.class.php
@@ -103,7 +103,7 @@ class WCF {
 		self::$autoloadDirectories['wcf'] = WCF_DIR . 'lib/';
 		
 		// define tmp directory
-		if (!defined('TMP_DIR')) define('TMP_DIR', util\BasicFileUtil::getTempFolder());
+		if (!defined('TMP_DIR')) define('TMP_DIR', util\FileUtil::getTempFolder());
 		
 		// start initialization
 		$this->initMagicQuotes();
